Refactor some test utilities into its own crate
Refactor some test utilities into its own crate.
The test-utils crate includes for now these functions:
- produce_macro_blocks
- fill_micro_blocks
- sign_macro_block
- sign_view_change
