+++ Cargo.toml
@@ -42,6 +42,7 @@ members = [
   "rpc-interface",
   "rpc-server",
   "tendermint",
+  "test-utils",
   "tools",
   "transaction-builder",
   "utils",
