+++ blockchain/Cargo.toml
@@ -44,6 +44,7 @@ atomic = "0.4"
 
 nimiq-block-production = { path = "../block-production", features = ["test-utils"] }
 nimiq-nano-primitives = { path= "../nano-primitives" }
+nimiq-test-utils = { path= "../test-utils" }
 
 [features]
 metrics = []
