+++ blockchain/tests/history_sync.rs
@@ -1,130 +1,16 @@
 use std::sync::Arc;
 
 use beserial::Deserialize;
-use nimiq_block::{
-    Block, MacroBlock, MacroBody, MultiSignature, TendermintIdentifier, TendermintProof,
-    TendermintProposal, TendermintStep, TendermintVote,
-};
 use nimiq_block_production::BlockProducer;
-use nimiq_blockchain::{AbstractBlockchain, Blockchain, PushResult};
-use nimiq_bls::{AggregateSignature, KeyPair, SecretKey};
-use nimiq_collections::bitset::BitSet;
+use nimiq_blockchain::{Blockchain, PushResult};
+use nimiq_bls::{KeyPair, SecretKey};
 use nimiq_database::volatile::VolatileEnvironment;
 use nimiq_genesis::NetworkId;
-use nimiq_hash::{Blake2bHash, Blake2sHash, Hash};
-use nimiq_primitives::policy;
 use nimiq_primitives::policy::{BATCHES_PER_EPOCH, BATCH_LENGTH, EPOCH_LENGTH};
+use nimiq_test_utils::blockchain::produce_macro_blocks;
 
 // Secret key of validator. Tests run with `genesis/src/genesis/unit-albatross.toml`
 const SECRET_KEY: &str = "196ffdb1a8acc7cbd76a251aeac0600a1d68b3aba1eba823b5e4dc5dbdcdc730afa752c05ab4f6ef8518384ad514f403c5a088a22b17bf1bc14f8ff8decc2a512c0a200f68d7bdf5a319b30356fe8d1d75ef510aed7a8660968c216c328a0000";
-
-// Produces a series of macro blocks (and the corresponding batches).
-fn produce_macro_blocks(num_macro: usize, producer: &BlockProducer, blockchain: &Arc<Blockchain>) {
-    for _ in 0..num_macro {
-        fill_micro_blocks(producer, blockchain);
-
-        let slots = blockchain.current_validators().unwrap();
-
-        let next_block_height = (blockchain.block_number() + 1) as u64;
-
-        let macro_block_proposal = producer.next_macro_block_proposal(
-            blockchain.time.now() + next_block_height * 1000,
-            0u32,
-            vec![],
-        );
-
-        let block = sign_macro_block(
-            TendermintProposal {
-                value: macro_block_proposal.header,
-                valid_round: None,
-            },
-            macro_block_proposal.body.unwrap(),
-            MacroBlock::create_pk_tree_root(&slots),
-        );
-
-        assert_eq!(
-            blockchain.push(Block::Macro(block)),
-            Ok(PushResult::Extended)
-        );
-    }
-}
-
-// Fill batch with micro blocks.
-fn fill_micro_blocks(producer: &BlockProducer, blockchain: &Arc<Blockchain>) {
-    let init_height = blockchain.block_number();
-
-    assert!(policy::is_macro_block_at(init_height));
-
-    let macro_block_number = init_height + BATCH_LENGTH;
-
-    for i in (init_height + 1)..macro_block_number {
-        let last_micro_block = producer.next_micro_block(
-            blockchain.time.now() + i as u64 * 1000,
-            0,
-            None,
-            vec![],
-            vec![0x42],
-        );
-
-        assert_eq!(
-            blockchain.push(Block::Micro(last_micro_block)),
-            Ok(PushResult::Extended)
-        );
-    }
-
-    assert_eq!(blockchain.block_number(), macro_block_number - 1);
-}
-
-// Signs a macro block proposal.
-fn sign_macro_block(
-    proposal: TendermintProposal,
-    extrinsics: MacroBody,
-    validator_merkle_root: Vec<u8>,
-) -> MacroBlock {
-    let keypair =
-        KeyPair::from(SecretKey::deserialize_from_vec(&hex::decode(SECRET_KEY).unwrap()).unwrap());
-
-    // Create a TendermintVote instance out of known properties.
-    // round_number is for now fixed at 0 for tests, but it could be anything,
-    // as long as the TendermintProof further down this function does use the same round_number.
-    let vote = TendermintVote {
-        proposal_hash: Some(proposal.value.hash::<Blake2bHash>()),
-        id: TendermintIdentifier {
-            block_number: proposal.value.block_number,
-            step: TendermintStep::PreCommit,
-            round_number: 0,
-        },
-        validator_merkle_root,
-    };
-
-    // create the to-be-signed hash
-    let message_hash = vote.hash::<Blake2sHash>();
-
-    // sign the hash
-    let signature = AggregateSignature::from_signatures(&[keypair
-        .secret_key
-        .sign_hash(message_hash)
-        .multiply(policy::SLOTS)]);
-
-    // create and populate signers BitSet.
-    let mut signers = BitSet::new();
-
-    for i in 0..policy::SLOTS {
-        signers.insert(i as usize);
-    }
-
-    // create the TendermintProof
-    let justification = Some(TendermintProof {
-        round: 0,
-        sig: MultiSignature::new(signature, signers),
-    });
-
-    MacroBlock {
-        header: proposal.value,
-        justification,
-        body: Some(extrinsics),
-    }
-}
 
 // Tests if the basic history sync works. It will try to push a succession of election and checkpoint
 // blocks. It does test if election blocks can be pushed after checkpoint blocks and vice-versa. It
