+++ test-utils/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "nimiq-test-utils"
+version = "0.1.0"
+authors = ["The Nimiq Core Development Team <info@nimiq.com>"]
+edition = "2018"
+description = "Test utilities for Nimiq's Rust implementation"
+homepage = "https://nimiq.com"
+repository = "https://github.com/nimiq/core-rs-albatross"
+license = "Apache-2.0"
+categories = ["cryptography::cryptocurrencies"]
+keywords = ["nimiq", "cryptocurrency", "blockchain"]
+
+[dependencies]
+hex = "0.4"
+
+beserial = { path = "../beserial" }
+nimiq-block = { path = "../primitives/block" }
+nimiq-blockchain = { path = "../blockchain" }
+nimiq-block-production = { path = "../block-production" }
+nimiq-bls = { path = "../bls" }
+nimiq-collections = { path = "../collections" }
+nimiq-hash = { path = "../hash" }
+nimiq-nano-primitives = { path = "../nano-primitives" }
+nimiq-primitives = { path = "../primitives" }
+nimiq-vrf = { path = "../vrf" }
